requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
pydantic==1.9.0
scikit-learn==1.0.2
azure-storage-blob==12.11.0